pyproject.toml
[project]
name = "tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.11.0",
    "fastapi>=0.115.0",
    "matplotlib>=3.10.7",
    "onnx>=1.19.1",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.12.0.88",
    "pillow>=11.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.17",
    "streamlit>=1.40.0",
    "timm>=1.0.22",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.47.0",
    "ultralytics>=8.3.225",
    "uvicorn[standard]>=0.32.0",
]
